requirement.txt
python-telegram-bot==21.5
yt-dlp==2024.8.6
